Rename document module to text_document
