+++ crates/ruff_server/src/edit/text_document.rs
